+++ src/main/java/com/buuz135/industrial/item/addon/AdultFilterAddonItem.java
@@ -1,6 +1,7 @@
 package com.buuz135.industrial.item.addon;
 
 import com.buuz135.industrial.proxy.ItemRegistry;
+import com.buuz135.industrial.tile.IAcceptsAdultFilter;
 import com.buuz135.industrial.tile.agriculture.AnimalIndependenceSelectorTile;
 import com.buuz135.industrial.utils.RecipeUtils;
 import net.minecraft.item.ItemStack;
@@ -14,7 +15,7 @@ public class AdultFilterAddonItem extends CustomAddon {
 
     @Override
     public boolean canBeAddedTo(SidedTileEntity machine) {
-        return machine instanceof AnimalIndependenceSelectorTile && !((AnimalIndependenceSelectorTile) machine).hasAdultFilterAddon();
+        return machine instanceof IAcceptsAdultFilter && !((IAcceptsAdultFilter) machine).hasAddon();
     }
 
     public void createRecipe() {