+++ src/main/java/com/buuz135/industrial/tile/IAcceptsAdultFilter.java
@@ -1,0 +1,6 @@
+package com.buuz135.industrial.tile;
+
+public interface IAcceptsAdultFilter {
+
+    boolean hasAddon();
+}