+++ src/main/java/com/buuz135/industrial/tile/mob/MobRelocatorTile.java
@@ -1,12 +1,15 @@
 package com.buuz135.industrial.tile.mob;
 
 import com.buuz135.industrial.IndustrialForegoing;
+import com.buuz135.industrial.item.addon.AdultFilterAddonItem;
 import com.buuz135.industrial.proxy.FluidsRegistry;
 import com.buuz135.industrial.tile.CustomColoredItemHandler;
+import com.buuz135.industrial.tile.IAcceptsAdultFilter;
 import com.buuz135.industrial.tile.WorkingAreaElectricMachine;
 import com.buuz135.industrial.tile.block.MobRelocatorBlock;
 import com.buuz135.industrial.utils.ItemStackUtils;
 import com.buuz135.industrial.utils.WorkUtils;
+import net.minecraft.entity.EntityAgeable;
 import net.minecraft.entity.EntityLiving;
 import net.minecraft.entity.item.EntityItem;
 import net.minecraft.entity.item.EntityXPOrb;
@@ -24,8 +27,9 @@ import net.minecraftforge.items.ItemStackHandler;
 import net.ndrei.teslacorelib.inventory.BoundingRectangle;
 
 import java.util.List;
+import java.util.concurrent.atomic.AtomicBoolean;
 
-public class MobRelocatorTile extends WorkingAreaElectricMachine {
+public class MobRelocatorTile extends WorkingAreaElectricMachine implements IAcceptsAdultFilter{
 
     private IFluidTank outExp;
     private ItemStackHandler outItems;
@@ -78,7 +82,11 @@ public class MobRelocatorTile extends WorkingAreaElectricMachine {
         List<EntityLiving> mobs = this.getWorld().getEntitiesWithinAABB(EntityLiving.class, area);
         if (mobs.size() == 0) return 0;
         FakePlayer player = IndustrialForegoing.getFakePlayer(world);
-        mobs.forEach(entityLiving -> entityLiving.attackEntityFrom(DamageSource.causePlayerDamage(player), Integer.MAX_VALUE));
+        AtomicBoolean hasWorked = new AtomicBoolean(false);
+        mobs.stream().filter(entityLiving -> !hasAddon() || (!(entityLiving instanceof EntityAgeable) || !entityLiving.isChild())).forEach(entityLiving -> {
+            entityLiving.attackEntityFrom(DamageSource.causePlayerDamage(player), Integer.MAX_VALUE);
+            hasWorked.set(true);
+        });
         List<EntityItem> items = this.getWorld().getEntitiesWithinAABB(EntityItem.class, area);
         for (EntityItem item : items) {
             if (!item.getItem().isEmpty()) {
@@ -86,7 +94,7 @@ public class MobRelocatorTile extends WorkingAreaElectricMachine {
                 item.setDead();
             }
         }
-        return 1;
+        return hasWorked.get() ? 1 : 0;
     }
 
 
@@ -105,5 +113,10 @@ public class MobRelocatorTile extends WorkingAreaElectricMachine {
     @Override
     protected void processFluidItems(ItemStackHandler fluidItems) {
         ItemStackUtils.processFluidItems(fluidItems, outExp);
+    }
+
+    @Override
+    public boolean hasAddon() {
+        return this.hasAddon(AdultFilterAddonItem.class);
     }
 }